Changed where Sentry is initialized.
